# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    API specification. Requests to an unknown route under `/api` return a 404
    response using the `Error` schema with the message `API route not found`.
    Requests with malformed JSON bodies return a 400 response using the `Error`
    schema with the message `Malformed JSON request body`. Unexpected server
    failures return a generic 500 response using the same schema. JSON request
    bodies that exceed the parser's size limit return a 413 response using the
    `Error` schema with the message `Request body too large`.
servers:
  - url: /api
    description: Base API path
...
paths:
  /admin/users:
    get:
      tags: [admin]
      operationId: listAdminUsers
      summary: List accounts for administrator role management
      responses:
        "200":
          description: User accounts and current roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/users/{id}/role:
    patch:
      tags: [admin]
      operationId: updateInstructorAccess
      summary: Grant or revoke a user's instructor role
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstructorAccessUpdate"
      responses:
        "200":
          description: Updated account role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Invalid role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator role or same-origin request required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/users/{id}/admin-access:
    post:
      tags: [admin]
      operationId: grantAdministratorAccess
      summary: Grant administrator access to an existing account
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministratorAccessGrantConfirmation"
      responses:
        "200":
          description: Account after administrator access was granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Explicit confirmation is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator role or same-origin request required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The account already has administrator access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [admin]
      operationId: revokeAdministratorAccess
      summary: Revoke administrator access from an account
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Account after administrator access was revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: The acting administrator cannot revoke their own access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator role or same-origin request required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revocation would remove the last administrator or bootstrap access is still configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/access-audit:
    get:
      tags: [admin]
      operationId: listInstructorAccessAudit
      summary: Search account role-change history
      parameters:
        - $ref: "#/components/parameters/AccessAuditAccount"
        - $ref: "#/components/parameters/AccessAuditActor"
      responses:
        "200":
          description: Matching role-change history, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstructorAccessAuditEntry"
        "403":
          description: Administrator role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          description: Invalid account or actor filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    AccessAuditAccount:
      name: account
      in: query
      required: false
      description: Case-insensitive substring matched against the changed account name, email, or ID.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    AccessAuditActor:
      name: actor
      in: query
      required: false
      description: Case-insensitive substring matched against the acting administrator name, email, or ID.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  schemas:
...
    AdminUser:
      type: object
      required: [id, email, firstName, lastName, role, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/AuthUserRole"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthUserRole:
      type: string
      enum: [learner, instructor, admin]
    InstructorAccessRole:
      type: string
      enum: [learner, instructor]
    InstructorAccessUpdate:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/InstructorAccessRole"
    AdministratorAccessGrantConfirmation:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean
          const: true
          description: Explicit confirmation that administrator access should be granted.
    InstructorAccessAuditEntry:
      type: object
      required:
        - id
        - actorId
        - actorEmail
        - actorName
        - targetId
        - targetEmail
        - targetName
        - previousRole
        - newRole
        - changedAt
      properties:
        id:
          type: string
        actorId:
          type: string
        actorEmail:
          type: ["string", "null"]
          format: email
        actorName:
          type: ["string", "null"]
        targetId:
          type: string
        targetEmail:
          type: ["string", "null"]
          format: email
        targetName:
          type: ["string", "null"]
        previousRole:
          $ref: "#/components/schemas/AuthUserRole"
        newRole:
          $ref: "#/components/schemas/AuthUserRole"
        changedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string